requirements.txt
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-google-genai==2.0.11
openai==1.102.0
python-dotenv==1.1.1
rich==14.1.0
pytest==8.3.4

